Move the operator extensions into their own headers
