+++ framework/src/mfem/integrators/NLCurlCurlIntegrator.C
@@ -11,7 +11,12 @@
 #ifdef MOOSE_MFEM_ENABLED
 
 #include "NLCurlCurlIntegrator.h"
+
+// Silence warning in the CI caused by this header
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wunused-parameter"
 #include "mfem/fem/integ/bilininteg_hcurl_kernels.hpp"
+#pragma GCC diagnostic pop
 
 namespace Moose::MFEM
 {
